+++ django_gcp/storage/gcloud.py
@@ -335,11 +335,14 @@ class GoogleCloudStorage(CompressStorageMixin, Storage):  # pylint: disable=abst
 
 
 class GoogleCloudMediaStorage(GoogleCloudStorage):  # pylint: disable=abstract-method
-    """Storage whose bucket name is taken from the GCP_STORAGE_MEDIA_NAME setting
+    """Storage for media files, configured via STORAGES['default']
+
+    This should be used as the BACKEND for the 'default' storage alias in Django's
+    STORAGES setting. Configuration options are passed via STORAGES['default']['OPTIONS'].
 
     This actually behaves exactly as a default instantiation of the base
-    ``GoogleCloudStorage`` class, but is there to make configuration more
-    explicit for first-timers.
+    ``GoogleCloudStorage`` class with store_key='media', but is there to make
+    configuration more explicit for first-timers.
 
     """
 
@@ -350,9 +353,12 @@ class GoogleCloudMediaStorage(GoogleCloudStorage):  # pylint: disable=abstract-m
 
 
 class GoogleCloudStaticStorage(GoogleCloudStorage):  # pylint: disable=abstract-method
-    """Storage defined with an appended bucket name (called "<bucket>-static")
+    """Storage for static files, configured via STORAGES['staticfiles']
 
-    We define that static files are stored in a different bucket than the (private) media files, which:
+    This should be used as the BACKEND for the 'staticfiles' storage alias in Django's
+    STORAGES setting. Configuration options are passed via STORAGES['staticfiles']['OPTIONS'].
+
+    We recommend storing static files in a different bucket than (private) media files, which:
         1. gives us less risk of accidentally setting private files as public
         2. allows us easier visual inspection in the console of what's private and what's public static
         3. allows us to set blanket public ACLs on the static bucket